Performed Player, Frame. New GamePlay, GameRules
Created the new Classes
-GamePlay: JFrame where the game is goint to played
-GameRules: recive the list of plays and return the frames with the game
rules after the processed data of the fle is given, but appling de
Bowling game rules.
. Ended the performence of the classes:
-Player: decided to hav an array of Frames where the player game
processes is introduced
-Frame: changed name form Frames. Now it indicates a single Frame so
every Frame have the atribute of being the final Frame, so the
programmer can indicate when is used and how to be used with a thirdPlay
score that's only available if the Frame is the FinalFrame.